+++ stencil-workspace/src/components/modus-date-picker/utils/modus-date-picker.types.tsx
@@ -3,8 +3,9 @@ import { ModusDateInputCustomEvent } from '../../../components';
 export type DateInputType = 'start' | 'end' | 'single';
 
 export type DateInputEventData = {
-  date: Date;
+  value: string;
   type: DateInputType;
+  inputString: string;
 };
 
 export type DateInputEvent = ModusDateInputCustomEvent<DateInputEventData>;